requirements.txt
pytest==7.4.3
requests==2.31.0
python-dotenv==1.0.0
pytest-html==4.1.1 